third_party/blink/web_tests/external/wpt/paint-timing/fcp-only/fcp-gradient-expected.txt

This is a testharness.js-based test.
FAIL Gradients should not count as contentful assert_equals: First contentful paint marked too early.  expected 0 but got 1
Harness: the test ran to completion.
